backend/requirements.txt
fastapi
uvicorn[standard]
aiortc
google-cloud-speech
google-cloud-translate
google-cloud-texttospeech